+++ CHANGELOG.md
@@ -10,6 +10,11 @@ Data Layer Observer follows semantic versioning when releasing updates.
 
 ## History
 
+### 1.7.0
+
+- Updated `beforeDestination` to accept a list allowing multiple operators to execute
+- Updated `convert` operator to accept negative `index` (read from end of a data list)
+
 ### 1.6.5
 
 - Adjusted suffixing behavior for `pageName`, `displayName`, and `email` to support FS APIs